# GH-117586: Speed up `pathlib.Path.glob()` by working with strings