--- a/Misc/NEWS.d/next/Library/2024-04-06-20-31-09.gh-issue-117586.UgWdRK.rst
+++ b/Misc/NEWS.d/next/Library/2024-04-06-20-31-09.gh-issue-117586.UgWdRK.rst
@@ -1,0 +1,1 @@
+Speed up :meth:`pathlib.Path.glob` by working with strings internally.