… Lib/pathlib/_glob.py