@barneygale 2024-04-06
This is the first PR in a series that will hopefully unify the globbing implementations in the `pathlib` and `glob` modules, and speed both up in the process.