@barneygale 2024-04-07
Hey @serhiy-storchaka, does this PR look alright to you? Not requesting a detailed review, more of a sanity check, given you've looked after the `glob` module for the last few years.

This PR doesn't affect `glob()` or `iglob()`, but it does move pathlib's globbing implementation into `glob.py`.

Thank you.