@barneygale 2024-04-10
I'll merge this now as it's important for #115060, which I'm hoping to get done in time for 3.13 beta 1.

But I'll leave `glob.glob()` and `glob.iglob()` unchanged in 3.13; any PRs I make will target 3.14.